upwork-automation/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
setuptools
aiohttp
requests
pandas
numpy
lxml  
groq
tiktoken 
faiss-cpu 
numpy 
langchain-text-splitters
langgraph
openai
langchain-openai



